upd: migrate to float, check withdraws borders

+++ internal/models/wallet.go
@@ -1,8 +1,8 @@
 package models
 
 type Wallet struct {
-	ID      int    `json:"id"`
-	UserID  int    `json:"user_id"`
-	Token   string `json:"token"`
-	Deposit int    `json:"deposit"`
+	ID      int     `json:"id"`
+	UserID  int     `json:"user_id"`
+	Token   string  `json:"token"`
+	Deposit float64 `json:"deposit"`
 }
